pyproject.toml
[project]
name = "estate-intelligence-backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "alembic>=1.18.4",
    "asyncpg>=0.30.0",
    "dash>=4.0.0",
    "fastapi>=0.135.1",
    "geopandas>=1.1.2",
    "greenlet>=3.3.2",
    "numpy>=2.4.2",
    "opentelemetry-api>=1.40.0",
    "pandas>=3.0.1",
    "plotly>=6.6.0",
    "pydantic-settings>=2.13.1",
    "scipy>=1.17.1",
    "shapely>=2.1.2",
    "sqlalchemy>=2.0.48",
    "structlog>=25.5.0",
    "uuid7>=0.1.0",
]

[dependency-groups]
dev = [
    "mypy>=1.19.1",
    "pre-commit>=4.5.1",
    "pytest>=9.0.2",
    "pytest-cov>=6.0.0",
    "pytest-asyncio>=0.25.3",
    "httpx>=0.28.1",
    "ruff>=0.15.4",
    "asgi-lifespan>=2.1.0",
    "aiosqlite>=0.22.1",
]

[tool.ruff]
target-version = "py311"
line-length = 100
[tool.ruff.lint]
select = [
    "E",      # pycodestyle errors
    "W",      # pycodestyle warnings
    "F",      # pyflakes
    "I",      # isort
    "N",      # pep8-naming
    "UP",     # pyupgrade
    "B",      # flake8-bugbear
    "C4",     # flake8-comprehensions
    "SIM",    # flake8-simplify
    "TCH",    # flake8-type-checking
]
ignore = [
    "E501",   # line too long (handled by formatter)
]
[tool.ruff.lint.isort]
known-first-party = ["estate-intelligence-backend"]
[tool.mypy]
python_version = "3.12"
strict = true
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_any_generics = true
check_untyped_defs = true
no_implicit_reexport = true
warn_redundant_casts = true
warn_unused_ignores = true
